requirements.txt
streamlit==1.25.0
tensorflow==2.13.0
pillow==9.0.1
numpy==1.21.2
